Fixing mariadb server config

+++ tests/ci/bootstrap.sh
@@ -62,7 +62,27 @@ then
 
     yum -y install ~/rpmbuild/RPMS/*/*.rpm
     mysql_install_db --user mysql
-    mv /etc/my.cnf.d/mariadb-server.cnf.rpmsave /etc/my.cnf.d/mariadb-server.cnf
+
+    if [ -f /etc/my.cnf.d/mariadb-server.cnf.rpmsave ]; then
+        mv /etc/my.cnf.d/mariadb-server.cnf.rpmsave /etc/my.cnf.d/mariadb-server.cnf
+    fi
+    if [ -f /etc/my.cnf.d/mariadb-server.cnf ]; then
+        >/etc/my.cnf.d/mariadb-server.cnf
+        echo "# this is read by the standalone daemon and embedded servers
+              [server]
+              sql_mode=
+              # this is only for the mysqld standalone daemon
+              # Settings user and group are ignored when systemd is used.
+              # If you need to run mysqld under a different user or group,
+              # customize your systemd unit file for mysqld/mariadb according to the
+              # instructions in http://fedoraproject.org/wiki/Systemd
+              [mysqld]
+              datadir=/var/lib/mysql
+              socket=/var/lib/mysql/mysql.sock
+              log-error=/var/log/mariadb/mariadb.log
+              pid-file=/run/mariadb/mariadb.pid" > /etc/my.cnf.d/mariadb-server.cnf
+    fi
+
     copy_template_httpd_conf
     ~/bin/services start
     mysql -e "CREATE USER 'root'@'gateway' IDENTIFIED BY '';
